Mark failed TCP connection attempts as errors
A failed TCP connection attempt attached its exception to the activity but left
the status Unset. Tracing backends treat an unset status as success, so failed
attempts did not show up in error-rate queries and the exception event was easy
to miss.

Pair AddException with SetStatus(ActivityStatusCode.Error), matching what
ConnectionFactory.CreateConnectionAsync and AutorecoveringConnection.Recovery
already do.

The parent "connection attempt" activity is deliberately left Unset when a later
endpoint succeeds: the overall operation did succeed, and failing over across
multiple endpoints is expected behavior.

The tests covering this live in SequentialIntegration rather than Integration.
ActivityListener is process-global, but Integration.csproj leaves the
CollectionPerAssembly attribute commented out, so its tests run in parallel.
Every Integration test class opens a connection in InitializeAsync, and each of
those fires its own "connection attempt" activity, so an unfiltered
ActivityRecorder counts other tests' activities as well as its own - which made
these tests fail intermittently on the integration-win32 leg, where the slower
CI runtime widens the window.

Scoping the recorders with a manually-started root activity and filtering by
TraceId or by parent identity does not solve this, and adds a dependency on
Activity.DefaultIdFormat, which is Hierarchical on net472 and W3C on net8.0.
Running serially, next to the other tracing tests, needs no filtering at all.
No ActivityRecorder usage is left in the parallel Integration project, which is
the invariant that keeps this from recurring.

The failover test installs an order-preserving IEndpointResolver because
DefaultEndpointResolver shuffles the endpoint list on every call, which would
otherwise let the reachable endpoint be tried first and produce a single TCP
attempt.
